feat: add video preview wrapper
